+++ app/src/main/java/fr/free/nrw/commons/MediaDataExtractor.java
@@ -23,6 +23,7 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
 import fr.free.nrw.commons.location.LatLng;
+import fr.free.nrw.commons.media.Media;
 import fr.free.nrw.commons.mwapi.MediaResult;
 import fr.free.nrw.commons.mwapi.MediaWikiApi;
 import timber.log.Timber;
