+++ app/src/main/java/fr/free/nrw/commons/bookmarks/pictures/BookmarkPicturesController.java
@@ -6,7 +6,7 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import fr.free.nrw.commons.Media;
+import fr.free.nrw.commons.media.Media;
 import fr.free.nrw.commons.bookmarks.Bookmark;
 import fr.free.nrw.commons.mwapi.MediaWikiApi;
