+++ app/src/main/java/fr/free/nrw/commons/category/CategoryImageController.java
@@ -5,7 +5,7 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import fr.free.nrw.commons.Media;
+import fr.free.nrw.commons.media.Media;
 import fr.free.nrw.commons.mwapi.MediaWikiApi;
 
 @Singleton
