+++ app/src/main/java/fr/free/nrw/commons/media/MediaDetailFragment.java
@@ -3,7 +3,6 @@ package fr.free.nrw.commons.media;
 import android.app.AlertDialog;
 import android.content.ClipData;
 import android.content.ClipboardManager;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.DataSetObserver;
 import android.net.Uri;
@@ -40,7 +39,6 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 import fr.free.nrw.commons.License;
 import fr.free.nrw.commons.LicenseList;
-import fr.free.nrw.commons.Media;
 import fr.free.nrw.commons.MediaDataExtractor;
 import fr.free.nrw.commons.MediaWikiImageView;
 import fr.free.nrw.commons.R;
