+++ app/src/test/kotlin/fr/free/nrw/commons/MediaTest.kt
@@ -1,5 +1,6 @@
 package fr.free.nrw.commons
 
+import fr.free.nrw.commons.media.Media
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
